requirements.txt
# absl-py==2.0.0
# altair==5.2.0
# appdirs==1.4.4
# asttokens==2.4.1
# astunparse==1.6.3
# attrs==23.1.0
# beautifulsoup4==4.12.2
# blinker==1.7.0
# CacheControl==0.13.1
# cachetools==5.3.2
# certifi==2023.11.17
# cffi==1.16.0
# charset-normalizer==3.3.2
# click==8.1.7
# cmdstanpy==1.2.0
# colorlover==0.3.0
# comm==0.2.0
# contourpy==1.2.0
# cryptography==41.0.7
# cufflinks==0.17.3
# cycler==0.12.1
# decorator==5.1.1
# exceptiongroup==1.2.0
# executing==2.0.1
# firebase-admin==6.3.0
# flatbuffers==23.5.26
# fonttools==4.46.0
# frozendict==2.3.10
# gast==0.5.4
# gitdb==4.0.11
# GitPython==3.1.40
# google-api-core==2.15.0
# google-api-python-client==2.111.0
# google-auth==2.25.2
# google-auth-httplib2==0.2.0
# google-auth-oauthlib==1.2.0
# google-cloud-core==2.4.1
# google-cloud-firestore==2.14.0
# google-cloud-storage==2.14.0
# google-crc32c==1.5.0
# google-pasta==0.2.0
# google-resumable-media==2.7.0
# googleapis-common-protos==1.62.0
# grpcio==1.60.0
# grpcio-status==1.60.0
# h5py==3.10.0
# holidays==0.38
# html5lib==1.1
# httplib2==0.22.0
# idna==3.6
# importlib-metadata==6.11.0
# importlib-resources==6.1.1
# ipython==8.18.1
# ipywidgets==8.1.1
# jedi==0.19.1
# Jinja2==3.1.2
# joblib==1.3.2
# jsonschema==4.20.0
# jsonschema-specifications==2023.11.2
# jupyterlab-widgets==3.0.9
# keras==2.15.0
# kiwisolver==1.4.5
# libclang==16.0.6
# lxml==4.9.3
# Markdown==3.5.1
# markdown-it-py==3.0.0
# MarkupSafe==2.1.3
# matplotlib==3.8.2
# matplotlib-inline==0.1.6
# mdurl==0.1.2
# ml-dtypes==0.2.0
# msgpack==1.0.7
# multitasking==0.0.11
# numpy==1.26.2
# oauthlib==3.2.2
# opt-einsum==3.3.0
# packaging==23.2
# pandas==2.1.4
# pandas-datareader==0.10.0
# parso==0.8.3
# peewee==3.17.0
# pexpect==4.9.0
# Pillow==10.1.0
# plotly==5.18.0
# prompt-toolkit==3.0.43
# prophet==1.1.5
# proto-plus==1.23.0
# protobuf==4.23.4
# ptyprocess==0.7.0
# pure-eval==0.2.2
# pyarrow==14.0.1
# pyasn1==0.5.1
# pyasn1-modules==0.3.0
# pycparser==2.21
# pydeck==0.8.1b0
# Pygments==2.17.2
# PyJWT==2.8.0
# pyparsing==3.1.1
# python-dateutil==2.8.2
# pytz==2023.3.post1
# referencing==0.32.0
# requests==2.31.0
# requests-oauthlib==1.3.1
# rich==13.7.0
# rpds-py==0.13.2
# rsa==4.9
# scikit-learn==1.3.2
# scipy==1.11.4
# six==1.16.0
# smmap==5.0.1
# soupsieve==2.5
# stack-data==0.6.3
# stanio==0.3.0
# streamlit==1.29.0
# streamlit-option-menu==0.3.6
# tenacity==8.2.3
# tensorboard==2.15.1
# tensorboard-data-server==0.7.2
# tensorflow==2.15.0
# tensorflow-estimator==2.15.0
# tensorflow-io-gcs-filesystem==0.34.0
# tensorflow-macos==2.15.0
# termcolor==2.4.0
# threadpoolctl==3.2.0
# toml==0.10.2
# toolz==0.12.0
# tornado==6.4
# tqdm==4.66.1
# traitlets==5.14.0
# typing_extensions==4.9.0
# tzdata==2023.3
# tzlocal==5.2
# uritemplate==4.1.1
# urllib3==2.1.0
# validators==0.22.0
# wcwidth==0.2.12
# webencodings==0.5.1
# Werkzeug==3.0.1
# widgetsnbextension==4.0.9
# wrapt==1.14.1
# yfinance==0.2.33
# zipp==3.17.0
absl-py==2.1.0
altair==5.2.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachecontrol==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.0
colorama==0.4.6
colorlover==0.3.0
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.1
cufflinks==0.17.3
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
firebase==4.0.1
firebase-admin==6.4.0
flatbuffers==23.5.26
fonttools==4.47.2
frozendict==2.4.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h5py==3.10.0
holidays==0.41
html5lib==1.1
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipython==8.18.1
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyterlab-widgets==3.0.9
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
multitasking==0.0.11
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pandas-datareader==0.10.0
parso==0.8.3
peewee==3.17.0
pillow==10.2.0
plotly==5.18.0
prompt-toolkit==3.0.43
prophet==1.1.5
proto-plus==1.23.0
protobuf==4.23.4
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sklearn==0.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
stanio==0.3.0
streamlit==1.30.0
streamlit-option-menu==0.3.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing-extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
webencodings==0.5.1
werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
yfinance==0.2.33
zipp==3.17.0